implement authorization

+++ Presentation/Controllers/ProfileController.cs
@@ -1,7 +1,8 @@
-﻿using Application.DTOs;
+﻿using System.Security.Claims;
 using Application.DTOs.Profile;
 using Application.Interfaces.Services;
 using Domain.Shared;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -9,6 +10,7 @@ namespace Presentation.Controllers;
 
 [Route("api/[controller]")]
 [ApiController]
+[Authorize]
 public class ProfileController : ControllerBase
 {
     private readonly IUserService _userService;
@@ -18,22 +20,27 @@ public class ProfileController : ControllerBase
         _userService = userService;
     }
 
-    [HttpGet("{id:guid}")]
-    public async Task<ActionResult> GetProfileAsync([FromRoute] Guid id)
+    [HttpGet]
+    public async Task<ActionResult> GetProfileAsync()
     {
-        if (ModelState.IsValid)
+        if (!ModelState.IsValid)
         {
-            var result = await _userService.GetUser(id);
-            return result.IsSuccess ? Ok(result.Value) : BadRequest(result.Error.description);
+            return BadRequest(Result
+                .Failure(new Error("Register request input", "Invalid")));
         }
 
-        return BadRequest(Result.
-            Failure(new Error("Register request input", "Invalid")));
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (!Guid.TryParse(userId, out Guid id))
+        {
+            return BadRequest("id fault");
+        }
+
+        var result = await _userService.GetUser(id);
+        return result.IsSuccess ? Ok(result.Value) : BadRequest(result.Error.description);
     }
 
-    [HttpPut("personal-information/{id:guid}")]
+    [HttpPut("personal-information/")]
     public async Task<ActionResult> UpdatePersonalInformationAsync(
-        [FromRoute] Guid id,
         UpdatePersonalInformationRequest request)
     {
         if (!ModelState.IsValid)
@@ -42,34 +49,38 @@ public class ProfileController : ControllerBase
                 .Failure(new Error("Register request input", "Invalid")));
         }
 
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (!Guid.TryParse(userId, out Guid id))
+        {
+            return BadRequest("id fault");
+        }
+
         var result = await _userService.UpdateProfileInformationAsync(id, request);
         return result.IsSuccess ? Ok(result.Value) : BadRequest(result.Error);
     }
 
-    [HttpPut("update-credentials/{id:guid}")]
+    [HttpPut("update-credentials/")]
     public async Task<ActionResult> UpdateCredentialsAsync(
-    [FromRoute] Guid id,
-    UpdateCredentialsRequest request)
+        UpdateCredentialsRequest request)
     {
         if (!ModelState.IsValid)
         {
             return BadRequest(Result
                 .Failure(new Error("Register request input", "Invalid")));
+        }
+
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (!Guid.TryParse(userId, out Guid id))
+        {
+            return BadRequest("id fault");
         }
 
         var result = await _userService.UpdateCredentialsAsync(id, request);
         return result.IsSuccess ? Ok() : BadRequest(result.Error);
     }
 
-    [HttpPut("update-photo/{id:guid}")]
-    public async Task<IActionResult> UpdatePhoto([FromRoute] Guid id, IFormFile file)
-    {
-        var result = await _userService.UpdateUserPhoto(id, file);
-        return result.IsSuccess ? Ok(result.Value): BadRequest(result.Error);
-    }
-
-    [HttpDelete("{id:guid}")]
-    public async Task<IActionResult> DeleteUser([FromRoute] Guid id)
+    [HttpPut("update-photo/")]
+    public async Task<IActionResult> UpdatePhoto(IFormFile file)
     {
         if (!ModelState.IsValid)
         {
@@ -77,6 +88,30 @@ public class ProfileController : ControllerBase
                 .Failure(new Error("Register request input", "Invalid")));
         }
 
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (!Guid.TryParse(userId, out Guid id))
+        {
+            return BadRequest("id fault");
+        }
+
+        var result = await _userService.UpdateUserPhoto(id, file);
+        return result.IsSuccess ? Ok(result.Value): BadRequest(result.Error);
+    }
+
+    [HttpDelete]
+    public async Task<IActionResult> DeleteUser()
+    {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(Result
+                .Failure(new Error("Register request input", "Invalid")));
+        }
+
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (!Guid.TryParse(userId, out Guid id))
+        {
+            return BadRequest("id fault");
+        }
 
         var result = await _userService.DeleteUserAsync(id);
         return result.IsSuccess ? Ok() : BadRequest(result.Error);
